Move files for testing to /tests directory
