--- tests/index.test.ts
@@ -1,3 +1,0 @@
-var context = require.context('../tests/', true, /.ts$/);
-context.keys().forEach(context);
-module.exports = context;
