+++ tests/src/index.test.ts
@@ -1,0 +1,3 @@
+var context = require.context('/tests/src/', true, /.ts$/);
+context.keys().forEach(context);
+module.exports = context;
